Deler opp godskriving omsorgspoeng i to forskjellige brev, dei er veldig ulike
