+++ pensjon-brevbaker-api-model/src/main/kotlin/no/nav/pensjon/brev/api/model/maler/Brevkode.kt
@@ -33,7 +33,8 @@ interface Brevkode {
 
     enum class Redigerbar : Brevkode {
         INFORMASJON_OM_SAKSBEHANDLINGSTID,
-        GODSKRIVING_AV_PENSJONSOPPTJENING_FOR_OMSORG_BARN,
+        GODSKRIVING_AV_PENSJONSOPPTJENING_FOR_OMSORG_BARN_FOER_1991,
+        GODSKRIVING_AV_PENSJONSOPPTJENING_FOR_OMSORG_BARN_ETTER_1991,
         PE_BEKREFTELSE_PAA_FLYKTNINGSTATUS,
         PE_FORESPOERSELOMDOKUMENTASJONAVBOTIDINORGE_ALDER,
         PE_FORESPOERSELOMDOKUMENTASJONAVBOTIDINORGE_ETTERLATTE,
