+++ pensjon-brevbaker/src/test/kotlin/no/nav/pensjon/brev/template/GenererAlleMaleneTest.kt
@@ -13,7 +13,7 @@ import org.junit.jupiter.params.provider.MethodSource
 
 class GenererAlleMaleneTest {
 
-    private val filterForPDF = listOf<String>()
+    private val filterForPDF = listOf<String>("godskriving")
 
     @Tag(TestTags.MANUAL_TEST)
     @ParameterizedTest(name = "{1}, {3}")
